+++ lib/interface.dart
@@ -1,8 +1,10 @@
-library dev.plugfox.ws.interface;
+library interface;
 
 export 'package:ws/src/client/message_stream.dart';
 export 'package:ws/src/client/metrics.dart';
 export 'package:ws/src/client/state.dart';
 export 'package:ws/src/client/status_codes.dart';
 export 'package:ws/src/client/web_socket_ready_state.dart';
+export 'package:ws/src/client/ws_client_fake.dart' show WebSocketClientFake;
 export 'package:ws/src/client/ws_client_interface.dart';
+export 'package:ws/src/client/ws_options.dart';